feat(signal): wire the statistical layer into the narrative + faithfulness eval [RNL-89]
compute_stats now runs Mann-Kendall, Sen slope, seasonal decomposition and a
3-month forecast, exposed as new (optional, additive) TrendStats fields. The
deterministic narrative reads like a statistician's: significance over bare
direction ("no significant trend, p=0.07"), seasonality with an honest
under-two-years caveat, and a projection with widening intervals.

allowed_from_stats now admits the inferential figures so an LLM narrative may
quote the p-value, slope/CI, seasonal strength and forecast and still pass the
faithfulness gate. Forecast values returned as native floats.

119 tests (2 new for the wiring); contract tests still hold (fields added, none
changed). On theft/Adelaide: MK p=0.075, seasonal strength 0.75 peaking March,
forecast 370/300/284.

+++ signalkit/analyst/stats.py
@@ -256,9 +256,9 @@ def forecast(
         points.append(
             ForecastPoint(
                 month=target_month,
-                point=round(max(0.0, point), 1),
-                lo=round(max(0.0, point - band), 1),
-                hi=round(point + band, 1),
+                point=float(round(max(0.0, point), 1)),
+                lo=float(round(max(0.0, point - band), 1)),
+                hi=float(round(point + band, 1)),
             )
         )
